Unnecessary float

+++ ann/ann_network.py
@@ -16,7 +16,7 @@ class Net(nn.Module):
     def forward(self,x):
         x = F.relu(self.fc1(x) )
         x = F.relu(self.fc2(x) )
-        x = self.fc3(x).float()
+        x = self.fc3(x)
         return F.log_softmax(x,dim=1)
 
 def LOG(message):
